[WIP] Feature/segment effort videos

+++ app/services/dropbox_sync.rb
@@ -1,12 +1,11 @@
 require 'dropbox_sdk'
 
 class DropboxSync
-
   def synchronize(dropbox_event)
-    client = DropboxClient.new( dropbox_event.user.access_token )
+    client = DropboxClient.new(dropbox_event.user.access_token)
     next_delta = true
     while next_delta
-      delta  = client.delta(dropbox_event.cursor, dropbox_event.path)
+      delta = client.delta(dropbox_event.cursor, dropbox_event.path)
       process_delta(dropbox_event, delta)
       next_delta = delta['has_more']
     end
@@ -14,9 +13,7 @@ class DropboxSync
 
   def process_delta(dropbox_event, delta)
     if dropbox_event.cursor != delta['cursor']
-      if delta['reset']
-        dropbox_event.dropbox_photos.destroy_all
-      end
+      dropbox_event.dropbox_photos.destroy_all if delta['reset']
       delta['entries'].each do |entry|
         process_delta_entry(entry)
       end
@@ -24,6 +21,9 @@ class DropboxSync
     end
   end
 
+  # rubocop:disable Metrics/MethodLength
+  # rubocop:disable Metrics/AbcSize
+  # rubocop:disable Metrics/PerceivedComplexity
   def process_delta_entry(entry)
     path = entry[0]
     metadata = entry[1]
@@ -31,8 +31,8 @@ class DropboxSync
     if !metadata
       Rails.logger.debug("Destroying photo with path #{path}")
       photo.destroy if photo.present?
-    elsif /\.((jpg)|(jpeg)|(png))$/ =~ path #is a photo
-      contents, contents_metadata = client.get_file_and_metadata(path, metadata['rev'])
+    elsif /\.((jpg)|(jpeg)|(png))$/ =~ path
+      contents = client.get_file_and_metadata(path, metadata['rev'])
       tmp = Tempfile.new 'image'
       begin
         tmp.binmode
@@ -42,17 +42,18 @@ class DropboxSync
         tmp.unlink
       end
       if photo
-        if photo.rev != metadata['rev'] #need to update
+        if photo.rev != metadata['rev'] # need to update
           photo.photo = tmp
           photo.rev = metadata['rev']
           photo.save
         end
       else
-        photo = DropboxPhoto.create(:path => path, :rev => metadata['rev'], :dropbox_event => dropbox_event)
+        photo = DropboxPhoto.create(path: path,
+                                    rev: metadata['rev'],
+                                    dropbox_event: dropbox_event)
         photo.photo = tmp
         photo.save
       end
     end
   end
-
 end
